Changed type of i to size_t

+++ 0x1D-search_algorithms/0-linear.c
@@ -10,13 +10,13 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int i = 0;
+	size_t i = 0;
 
 	if (array && size != 0)
 	{
 		for (; i < size; i++)
 		{
-			printf("Value checked array[%d] = [%d]\n", i, array[i]);
+			printf("Value checked array[%lu] = [%d]\n", i, array[i]);
 			if (array[i] == value)
 				return (i);
 		}
